requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jinja2==3.1.3
